+++ docs/CHANGELOG.md
@@ -47,6 +47,7 @@ it is attached to. ([#1282])
 
 ### Changed
 
+* `#[derive(PrefabData)]` now supports enums as well as structs
 * Make `frame_limiter::do_sleep` calculate the amount of time to sleep instead of calling `sleep(0)` ([#1446])
 * Make `application_root_dir` return a `Result<Path>` instead of a `String` ([#1213])
 * Remove unnecessary texture coordinates offset in `Sprite::from_pixel_values` ([#1267])